% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installSDK.R
\name{checkPathSetup}
\alias{checkPathSetup}
\title{Checks for existence of blangSDK Command Line Interface files}
\usage{
checkPathSetup(blangPath)
}
\description{
Checks for existence of blangSDK Command Line Interface files
}
