% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResult.R
\name{getResult.blangModel}
\alias{getResult.blangModel}
\title{Returns the samples of the given random variable as a dataframe.}
\usage{
getResult.blangModel(obj, variable, output.path = NULL)
}
\arguments{
\item{obj}{blangModel: the model from which the results are pulled}

\item{variable}{string: name of the random variable}

\item{output.path}{string: optional parameter to specify an output
location to copy the csv files to. By default set to
blangModel$out.loc (just a way to add another path without
having to change the default model out.loc).}
}
\value{
dataframe of the samples for the given random variable
}
\description{
Returns the samples of the given random variable as a dataframe.
}
