% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installSDK.R
\name{installSDK}
\alias{installSDK}
\title{Downloads current blangSDK repository from GitHub and places it in project folder.}
\usage{
installSDK(olddir, instPath)
}
\description{
Downloads current blangSDK repository from GitHub and places it in project folder.
}
