% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blangModel.R
\name{blangModel}
\alias{blangModel}
\title{Runs a given blang Model}
\usage{
blangModel(project.path = "", model.name = "", data = NULL,
  blang.args = paste("--model", model.name, sep = " "), out.loc = NULL,
  data.name = NULL)
}
\arguments{
\item{project.path}{filepath to the blang project root directory}

\item{model.name}{the name of the model to be run, as a string}

\item{data}{a dataframe the data passed into the model}

\item{blang.args}{a string of the runtime arguments to be used by blang}

\item{out.loc}{optional parameter to set the filepath of model result files}

\item{data.name}{optional string argument, set to the name of the
GlobalDataSource param in the blang model if known, otherwise
uses the name of the data param to pass into blang.}
}
\value{
--
}
\description{
Runs a given blang Model
}
