% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{run.blangModel}
\alias{run.blangModel}
\title{Run the given blang model}
\usage{
run.blangModel(obj, run.args = NULL)
}
\arguments{
\item{obj}{is a blangModel object, which will be run with the
arguments provided to it via obj$blang.args}

\item{run.args}{optional parameter that allows you to modify
the default blangModel$blang.args for this run. If NULL,
it will just use blangModel$blang.args.}
}
\value{
returns a blangModel object with this runs results
    path and arguments. After a run is complete, the list of
    sampled variables can be accessed via modelName$samples
    and the temperatures via modelName$temps.
}
\description{
Given the name of a blang model (name of .bl file), run
the model with no additional arguments.
}
